Money Long Change

+++ app/src/main/java/com/example/wallet/services/FirestoreService.kt
@@ -158,7 +158,7 @@ class FirestoreService {
                     // Crear el objeto Player con los datos del documento
                     Player(
                         name = document.getString("Name") ?: "Invitado",
-                        money = (document.getLong("Money") ?: 0L).toInt(),
+                        money = document.getLong("Money") ?: 0L,
                         uid = document.getString("Uid") ?: uid
                     )
                 }
